+++ src/modals.ts
@@ -49,26 +49,25 @@ export class LogsModal extends Modal {
             text: " Detection of multiple attachment reference links - logs ",
         });
         headerEl.addClass("modal-title");
-
-        // the numbe of referencing the same image in current note  more than once
-        if (this.state > 1) {
-            this.showPrompt();
-            this.showCloseBtn(this.getButtonWrapper(), this);
-        } else {
-            this.showLogs();
-            this.showCloseBtn(this.getButtonWrapper(), this);
-            this.showRemoveLinkBtn(this.getButtonWrapper(), this);
+        // if the deleted image referenced by mutilp notes
+        if (this.state === 1) this.showLogs();
+        const buttonWrapper = this.contentEl.createEl("div") as HTMLDivElement;
+        buttonWrapper.addClass("fast-image-cleaner-center-wrapper");
+        // if the deleted image referenced by mutilp notes
+        if (this.state === 1) {
+            this.showCloseBtn(buttonWrapper, this);
+            this.showRemoveLinkBtn(buttonWrapper, this);
         }
+        // the numbe of referencing the same image in current note  more than once
+        if (this.state === 2) {
+            this.showPrompt(this);
+        }
+
     }
     showLogs() {
         const logs = this.contentEl.createEl("div");
         logs.addClass("fast-image-cleaner-log");
         logs.setText(this.getLog());
-    }
-    getButtonWrapper(): HTMLDivElement {
-        const buttonWrapper = this.contentEl.createEl("div") as HTMLDivElement;
-        buttonWrapper.addClass("fast-image-cleaner-center-wrapper");
-        return buttonWrapper;
     }
     showCloseBtn(buttonWrapper: HTMLDivElement, myModal: Modal) {
         // --------closeButton--------
@@ -88,16 +87,24 @@ export class LogsModal extends Modal {
             "aria-label",
             "Continue to remove the reference link to the current attachment in the current document"
         );
+        removeLinkButton.addClass('mod-warning');
         removeLinkButton.addEventListener("click", () => {
             removeReferenceLink(this.FileBaseName, this.currentMd);
             myModal.close();
         });
     }
-    showPrompt() {
+    showPrompt(myModal: Modal) {
         const prompt = this.contentEl.createEl("span", {
             text: "Detected that the image you are attempting to delete is being referenced multiple times within the current document. \n As a result. We kindly ask that you manually remove the link.",
         });
         prompt.addClass("fast-image-cleaner-prompt");
-        this.showLogs();
+        const buttonWrapper = this.contentEl.createEl("div") as HTMLDivElement;
+        const closeButton = buttonWrapper.createEl("button", {
+            text: "close",
+        });
+        closeButton.setAttribute("aria-label", "close the window");
+        closeButton.addEventListener("click", () => {
+            myModal.close();
+        });
     }
 }