+++ src/utils/deleteNote.ts
@@ -3,14 +3,7 @@ import NathanDeletefile from "src/main";
 
 export const deleteNote = (
     file: TFile,
-    plugin: NathanDeletefile
 ) => {
-    const deleteOption = plugin.settings.deleteOption;
-    if (deleteOption === ".trash") {
-        app.vault.trash(file, false);
-    } else if (deleteOption === "system-trash") {
-        app.vault.trash(file, true);
-    } else if (deleteOption === "permanent") {
-        app.vault.delete(file);
-    }
+    // @ts-ignore
+    app.fileManager.promptForDeletion(file)
 };