+++ changelog.md
@@ -3,3 +3,4 @@
 2023 年 4 月 6 日
 
 -   ✨ feat：Delete all attachments referenced by the note when deleting current note
+-   ✨ feat: Use the public api `app.fileManager.promptForDeletion(file)` to handle file deletion